[BE] Refactor : 주어진 지도 크기에 맞게 응답한다. (#133)
* Refactor : 지도 가로세로 1km 로 수정

* Refactor : 섹터당 총 갯수 api 응답 수정

* bug: 전체 조회 api 안되는 문제 수정

* bug: 책 검색 생략가능하게 변경

* bug: 액세스토큰 못찾았을때 내는 예외 추가

* Refactor : 섹터를 1km 고정값이 아니라 주어진 가로세로에 맞게 응답하게 변경

+++ backend/src/main/java/com/dongnebook/domain/member/dto/response/MerchantSectorCountResponse.java
@@ -9,8 +9,9 @@ import lombok.Getter;
 public class MerchantSectorCountResponse {
 
 	private Long merchantCount;
-	private Long sector;
+	private Integer sector;
 	private Location location;
+
 
 	public void plusBookCount(){
 		if(merchantCount ==null){
@@ -23,7 +24,7 @@ public class MerchantSectorCountResponse {
 		this.location = location;
 	}
 
-	public void initSector(Long sector) {
+	public void initSector(Integer sector) {
 		this.sector = sector;
 	}
 }
